requirements.txt
streamlit
snowflake-connector-python
python-dotenv
pyarrow<10.1.0,>=10.0.1
pandas

yfinance